pyproject.toml
[tool.poetry]
name = "nr-scft-ml"
version = "0.1.0"
description = "The package tool for chi parameters prediction from SLD profile"
authors = ["Yuqing Qiao <qiao.yuqi@northeastern.edu>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.11.0.86"
torch = "^2.6.0"
numpy = "^2.1.0"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
scipy = "^1.15.1"
torchvision = "^0.21.0"
typer = "^0.15.1"
pyyaml = "^6.0.2"
llvmlite = ">=0.44.0"
numba = ">=0.61.0"
refl1d = { version = "^1.0.0a0", allow-prereleases = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

